+++ Components/NMOS.cpp
@@ -1,5 +1,5 @@
 #include "NMOS.h"
-NMOS::NMOS(string id, map<string, int>& info, map<string, string>& netlist):ThreeTerminal(id , netlist,info)
+NMOS::NMOS(string id, map<string, double>& info, map<string, string>& netlist):ThreeTerminal(id , netlist,info)
 {
 }
 
@@ -10,5 +10,5 @@ string NMOS::getType()
 
 void NMOS::printComponentInfo()
 {
-	printf("NMOS component with default value : %f and three terminals : %s %s %s", this->defaultValue, this->getT1().c_str(), this->getT2().c_str(), this->getT3().c_str());
+	printf("NMOS component with default value : %lf and three terminals : %s %s %s", this->defaultValue, this->getT1().c_str(), this->getT2().c_str(), this->getT3().c_str());
 }
