+++ Components/PMOS.cpp
@@ -1,5 +1,5 @@
 #include "PMOS.h"
-PMOS::PMOS(string id, map<string, int>& info, map<string, string>& netlist):ThreeTerminal(id , netlist,info)
+PMOS::PMOS(string id, map<string, double>& info, map<string, string>& netlist):ThreeTerminal(id , netlist,info)
 {
 }
 
@@ -10,5 +10,5 @@ string PMOS::getType()
 
 void PMOS::printComponentInfo()
 {
-	printf("PMOS component with default value : %f and three terminals : %s %s %s", this->defaultValue, this->getT1().c_str(), this->getT2().c_str(), this->getT3().c_str());
+	printf("PMOS component with default value : %lf and three terminals : %s %s %s", this->defaultValue, this->getT1().c_str(), this->getT2().c_str(), this->getT3().c_str());
 }
