+++ Components/VoltageSource.cpp
@@ -1,6 +1,6 @@
 #include "VoltageSource.h"
 
-VlotageSource::VlotageSource(string id, map<string, string>& VlotageSourceNetlist, map<string, int>& VlotageSourceValues):TwoTerminal(id , VlotageSourceNetlist , VlotageSourceValues)
+VlotageSource::VlotageSource(string id, map<string, string>& VlotageSourceNetlist, map<string, double>& VlotageSourceValues):TwoTerminal(id , VlotageSourceNetlist , VlotageSourceValues)
 {
 
 }
@@ -12,5 +12,5 @@ string VlotageSource::getType()
 
 void VlotageSource::printComponentInfo()
 {
-	printf("VlotageSource with default value : %d, t1 : %s, t2 : %s\n", this->defaultValue, this->getT1().c_str(), this->getT2().c_str());
+	printf("VlotageSource with default value : %lf, t1 : %s, t2 : %s\n", this->defaultValue, this->getT1().c_str(), this->getT2().c_str());
 }
